RELEASE 1.0
Full release of the game!
The final version of this project is finally here!

Now in just 3 weeks this project went from an old forgotten unity project with just the movement function and 3 levels to a fully playable game!

CHANGELOG from the last update:

- death counter

- pause menu

- fixed issues with UI on the main menu

- new key bindings (W = thrust, A; D = turn) NOTE: old binding still works!

CHANGELOG from how this game looked before i picked it back up:

-  more gameplay levels

-  main menu with level selection and settings functions

-  death counter

- pause menu